pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[project]
name = "histserv"
description = "Histogramming as a Service"
readme = "README.md"
authors = [
    { name = "pfackeldey", email = "fackeldey.peter@gmail.com" }
]
license = "BSD-3-Clause"
license-files = ["LICENSE"]
classifiers = [
    "Development Status :: 1 - Planning",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "Topic :: Scientific/Engineering :: Mathematics",
    "Topic :: Scientific/Engineering :: Physics",
    "Typing :: Typed",
]
keywords = [
    "histogram",
    "boost-histogram",
    "distributed",
    "serving",
]
dynamic = ["version"] # version is set in src/histserv/__init__.py
requires-python = ">=3.10"
dependencies = [
    "grpcio>=1.76.0",
    "hist>=2.9.0",
    "boost-histogram>=1.7.1",
    "uhi>=1.0.0",
    "numcodecs>=0.13.1",
    "grpcio-tools>=1.78.0",
]

[project.urls]
Homepage = "https://github.com/pfackeldey/histserv"
"Bug Tracker" = "https://github.com/pfackeldey/histserv/issues"
Discussions = "https://github.com/pfackeldey/histserv/discussions"

[dependency-groups]
dev = [
    "ipython",
    "h5py", # needed only for the `.flush` RPC call on the server
    "pytest>=8.0.0",
]

[project.scripts]
histserv = "histserv.__main__:run"

[tool.hatch]
version.path = "src/histserv/__init__.py"

[tool.hatch.build.targets.sdist]
exclude = [
    "/example",
]

[tool.hatch.build.targets.wheel]
packages = ["src/histserv"]
exclude = [
    "/example",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ty.src]
exclude = [
    "src/histserv/protos/*",
    "example/*",
    "tests/*",
]
